gprs: RoamingAllowed always
the property should be removed completely, but this would cause problems
of backward compatibility of the interface

therefore the property is set true all the time.

the whole property was a dirty patch up for avoiding extra costs on some
2G networks that were charging only for attaching on GPRS
the current behavior is more compatible with modern network,
especially LTE where the attach is part of the network access,
but also for updated legacy networks, that expect both registration
and attach at the same time.