Modem Signal Strength behavior touch-up
Disable the update of the signal strength from network
notification.
Modify the method to query the signal strength by invoking
the SMONI AT command.
Setup the new method to be invoked repeatdly just when the
modem is registred and notify when a new signal strength
is available.
Destroy the timeout and not call the new method again when
the modem is offline.

main author: khaled romdhani <khaled.romdhani@ardia.com.tn>